Restructure FP16 clamp patch into pipelines/z_image/patch_fp16.py subpackage
